Fixed ViewModels
The issue was that two of my projects had the same assembly name.
I had to fix that first.
After that, I found out that one of my accessed a null object.
object reference not set to an instance, means that in constructing the
object, it came across a null reference.
I had to follow all the constructors to find it.
It would be nice if it told you which object it was, since it has to
know (it did just try to access it)